# Faster dotnetcore, add parallel, better timing