--- a/README.md
+++ b/README.md
@@ -58,7 +58,7 @@ Install latest version of .NET Core
 
 Go to `dnx/`  
 `dotnet restore` (first time)  
-`dotnet run`
+`dotnet run --configuration Release`
 
 ### Haskell