--- a/dnx/Program.cs
+++ b/dnx/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Threading;
 using System.Threading.Tasks;
 
 
@@ -10,20 +12,68 @@ namespace ActorBenchmark
     {
         static void Main(string[] args)
         {
-            long limit = 1000000;
-            DateTime dt = DateTime.Now;
-            var x = skynetSync(0, limit, 10);
-            Console.WriteLine(x);
-            DateTime dt2 = DateTime.Now;
-            Console.WriteLine("Sync sec: {0:0.000}", (dt2 - dt).TotalSeconds);
-            var x2 = Task.Run(() => skynetAsync(0, limit, 10));
-            x2.Wait();
-            DateTime dt3 = DateTime.Now;
-            Console.WriteLine(x2.Result);
-            Console.WriteLine("Async sec: {0:0.000}", (dt3 - dt2).TotalSeconds);
-            //Console.ReadLine();
+            Console.WriteLine($"Arch {(IntPtr.Size == 8 ? "64 bit" : "32 bit")}");
+
+            var runs = 3;
+            for (var i = 0; i < runs; i++)
+            {
+                Run(i == runs - 1);
+            }
         }
-        static object taskLock = new object();
+
+        private static void Run(bool output)
+        {
+            long limit = 1000000;
+            var sw = Stopwatch.StartNew();
+
+            sw.Restart();
+            var x = skynetSync(0, limit, 10);
+            sw.Stop();
+            var t1 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x);
+                Console.WriteLine($"1 Thread - Sync: {t1:0.000}ms");
+            }
+            CleanUp();
+
+            sw.Restart();
+            var x2 = skynetAsync(0, limit, 10).Result;
+            sw.Stop();
+            var t2 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x2);
+                Console.WriteLine($"1 Thread - Async: {t2:0.000}ms");
+            }
+            CleanUp();
+
+            sw.Restart();
+            var x3 = skynetThreadpoolAsync(0, limit, 10).Result;
+            sw.Stop();
+            var t3 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x3);
+                Console.WriteLine($"Parallel Async: {t3:0.000}ms");
+            }
+            CleanUp();
+
+            sw.Restart();
+            var x4 = skynetParallel(0, limit, 10);
+            sw.Stop();
+            var t4 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x4);
+                Console.WriteLine($"Parallel Sync: {t4:0.000}ms");
+            }
+            CleanUp();
+        }
 
         private static long skynetSync(long num, long size, long div)
         {
@@ -34,7 +84,6 @@ namespace ActorBenchmark
             else
             {
                 long sum = 0;
-                var tasks = new List<Task<long>>();
                 for (var i = 0; i < div; i++)
                 {
                     var sub_num = num + i * (size / div);
@@ -44,15 +93,15 @@ namespace ActorBenchmark
             }
         }
 
-        private static Task<long> skynetAsync(long num, long size, long div)
+        private static Task<long> skynetThreadpoolAsync(long num, long size, long div)
         {
             if (size == 1)
             {
                 return Task.FromResult(num);
             }
             else
-            {                               
-                var tasks = new List<Task<long>>();
+            {
+                var tasks = new List<Task<long>>((int)div);
                 for (var i = 0; i < div; i++)
                 {
                     var sub_num = num + i * (size / div);
@@ -63,14 +112,77 @@ namespace ActorBenchmark
             }
         }
 
+        private static Task<long> skynetAsync(long num, long size, long div)
+        {
+            if (size == 1)
+            {
+                return Task.FromResult(num);
+            }
+            else
+            {
+                var tasks = new List<Task<long>>((int)div);
+                for (var i = 0; i < div; i++)
+                {
+                    var sub_num = num + i * (size / div);
+                    var task = skynetAsync(sub_num, size / div, div);
+                    tasks.Add(task);
+                }
+                return Task.WhenAll(tasks).ContinueWith(skynetAggregator);
+            }
+        }
+
         static long skynetAggregator(Task<long[]> children)
         {
             long sumAsync = 0;
-            foreach (var x in children.Result)
+            var results = children.Result;
+            for (var i = 0; i < results.Length; i++)
             {
-                sumAsync += x;
+                sumAsync += results[i];
             }
             return sumAsync;
+        }
+
+        private static long skynetParallel(long num, long size, long div)
+        {
+            if (size == 1)
+            {
+                return num;
+            }
+            else
+            {
+                long total = 0;
+
+                long[] source = new long[div];
+                for (var i = 0; i < div; i++)
+                {
+                    source[i] = i;
+                }
+
+                var rangePartitioner = Partitioner.Create(0L, source.Length);
+
+                Parallel.ForEach(rangePartitioner,
+                    () => 0L, 
+                    (range, loopState, runningtotal) =>
+                        {
+                            for (long i = range.Item1; i < range.Item2; i++)
+                            {
+                                var sub_num = num + i * (size / div);
+                                runningtotal += skynetSync(sub_num, size / div, div);
+                            }
+                            return runningtotal;
+                        }, 
+                    (subtotal) => Interlocked.Add(ref total, subtotal)
+                );
+
+                return total;
+            }
+        }
+
+        static void CleanUp()
+        {
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
+            GC.Collect();
         }
     }
 }