--- a/dnx/Program.cs
+++ b/dnx/Program.cs
@@ -12,7 +12,7 @@ namespace ActorBenchmark
     {
         static void Main(string[] args)
         {
-            Console.WriteLine($"Arch {(IntPtr.Size == 8 ? "64 bit" : "32 bit")}");
+            Console.WriteLine($"Arch {(IntPtr.Size == 8 ? "64 bit" : "32 bit")} - Cores {Environment.ProcessorCount}");
 
             var runs = 3;
             for (var i = 0; i < runs; i++)
@@ -51,6 +51,18 @@ namespace ActorBenchmark
             CleanUp();
 
             sw.Restart();
+            var x5 = skynetValueTaskAsync(0, limit, 10).Result;
+            sw.Stop();
+            var t5 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x5);
+                Console.WriteLine($"1 Thread - ValueTask Async: {t5:0.000}ms");
+            }
+            CleanUp();
+
+            sw.Restart();
             var x3 = skynetThreadpoolAsync(0, limit, 10).Result;
             sw.Stop();
             var t3 = sw.Elapsed.TotalMilliseconds;
@@ -59,6 +71,18 @@ namespace ActorBenchmark
             {
                 Console.WriteLine(x3);
                 Console.WriteLine($"Parallel Async: {t3:0.000}ms");
+            }
+            CleanUp();
+
+            sw.Restart();
+            var x6 = skynetThreadpoolValueTaskAsync(0, limit, 10).Result;
+            sw.Stop();
+            var t6 = sw.Elapsed.TotalMilliseconds;
+
+            if (output)
+            {
+                Console.WriteLine(x5);
+                Console.WriteLine($"Parallel - ValueTask Async: {t6:0.000}ms");
             }
             CleanUp();
 
@@ -73,6 +97,7 @@ namespace ActorBenchmark
                 Console.WriteLine($"Parallel Sync: {t4:0.000}ms");
             }
             CleanUp();
+
         }
 
         private static long skynetSync(long num, long size, long div)
@@ -140,6 +165,67 @@ namespace ActorBenchmark
                 sumAsync += results[i];
             }
             return sumAsync;
+        }
+
+
+        private static ValueTask<long> skynetValueTaskAsync(long num, long size, long div)
+        {
+            if (size == 1)
+            {
+                return num;
+            }
+            else
+            {
+                long subtotal = 0;
+                List<Task<long>> tasks = null;
+                
+                for (var i = 0; i < div; i++)
+                {
+                    var sub_num = num + i * (size / div);
+                    var task = skynetValueTaskAsync(sub_num, size / div, div);
+                    if (task.IsCompleted)
+                    {
+                        subtotal += task.Result;
+                    }
+                    else
+                    {
+                        if (tasks == null)
+                        {
+                            tasks = new List<Task<long>>((int)div);
+                        }
+                        tasks.Add(task.AsTask());
+                    }
+                }
+
+                if (tasks == null)
+                {
+                    return subtotal;
+                }
+                else if (subtotal > 0)
+                {
+                    tasks.Add(Task.FromResult(subtotal));
+                }
+                return Task.WhenAll(tasks).ContinueWith(skynetAggregator);
+            }
+        }
+
+        private static Task<long> skynetThreadpoolValueTaskAsync(long num, long size, long div)
+        {
+            if (size == 1)
+            {
+                return Task.FromResult(num);
+            }
+            else
+            {
+                var tasks = new List<Task<long>>((int)div);
+                for (var i = 0; i < div; i++)
+                {
+                    var sub_num = num + i * (size / div);
+                    var task = Task.Run(() => skynetValueTaskAsync(sub_num, size / div, div).AsTask());
+                    tasks.Add(task);
+                }
+                return Task.WhenAll(tasks).ContinueWith(skynetAggregator);
+            }
         }
 
         private static long skynetParallel(long num, long size, long div)