--- a/dnx/project.json
+++ b/dnx/project.json
@@ -4,10 +4,11 @@
         "emitEntryPoint": true
     },
 
-    "dependencies": {
-        "NETStandard.Library": "1.0.0-rc2-23811",
-        "System.Threading.Tasks.Parallel": "4.0.1-rc2-23805"
-    },
+  "dependencies": {
+    "NETStandard.Library": "1.0.0-rc2-23811",
+    "System.Threading.Tasks.Extensions": "4.0.0-rc2-23812",
+    "System.Threading.Tasks.Parallel": "4.0.1-rc2-23805"
+  },
 
     "frameworks": {
         "dnxcore50": { }