@benaadams 2016-02-15
Example output 
```
Arch 64 bit
499999500000
1 Thread - Sync: 44.023ms
499999500000
1 Thread - Async: 186.796ms
499999500000
Parallel Async: 90.184ms
499999500000
Parallel Sync: 13.680ms
```