@benaadams 2016-02-15
With the new ValueTask async
```
Arch 64 bit
499999500000
1 Thread - Sync: 42.482ms
499999500000
1 Thread - Async: 186.755ms
499999500000
1 Thread - ValueTask Async: 65.250ms
499999500000
Parallel Async: 86.657ms
499999500000
Parallel - ValueTask Async: 20.014ms
499999500000
Parallel Sync: 13.094ms
```