@benaadams 2016-02-15
Also if you actually compile it in release mode...
```
dotnet compile --configuration Release -o bin
bin\dnx
```
The results are
```
Arch 64 bit - Cores 4
499999500000
1 Thread - Sync: 5.756ms
499999500000
1 Thread - Async: 146.579ms
499999500000
1 Thread - ValueTask Async: 16.917ms
499999500000
Parallel Async: 86.191ms
499999500000
Parallel - ValueTask Async: 4.988ms
499999500000
Parallel Sync: 1.661ms
```