@Jetski5822 2016-02-15
whoah, didn't realize what ValueTask Async gave us! +1